+++ src/contexts/cookidoo/application/commands/shopping-list-remove-custom-recipe-ingredients/shopping-list-remove-custom-recipe-ingredients.command.ts
@@ -1,0 +1,8 @@
+/** Command: remove the ingredients of the given custom recipes from the list. */
+export class ShoppingListRemoveCustomRecipeIngredientsCommand {
+  public readonly recipeIds: string[];
+
+  constructor(input: { recipeIds: string[] }) {
+    this.recipeIds = input.recipeIds;
+  }
+}